% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks.plot.R
\name{cooks.plot}
\alias{cooks.plot}
\title{Plotting cook's distance for linear mixed effect models.}
\usage{
cooks.plot(mod, data, cut)
}
\arguments{
\item{mod}{a lmer model}

\item{data}{the dataset used in the object}

\item{cut}{cutoff value, unless specified, equals 4/n where n equals the observation number in the dataset.}
}
\description{
This function plots the cook's distance (calculated by HLMdiag ::cooks.distance) a linear mixed effect model fitted by lmer function.
}
\examples{
data(sleepstudy)
mod<-lmer(Reaction ~ Days + (Days|Subject), data=sleepstudy)
cooks.plot(mod, data=sleepstudy)
}
\author{
Yian Guo, Yuting Yang, Xiaoman Zhai
}
