% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev.plot.R
\name{lev.plot}
\alias{lev.plot}
\title{Plotting leverages against pearson residuals for linear mixed effect models}
\usage{
lev.plot(mod, type = "overall")
}
\arguments{
\item{mod}{a lmer model}

\item{type}{'overall' plots the overall leverage, 'fixef' plots the leverage corresponding to the fixed effects, 'ranef' plots the leverage corresponding to the random effects. Default is 'overall'.}
}
\description{
This function plots the leverages (calculated by HLMdiag ::leverage) against pearson residuals from a linear mixed effect model fitted by lmer function.
}
\examples{
data(sleepstudy)
mod<-lmer(Reaction ~ Days + (Days|Subject), data=sleepstudy)
lev.plot(mod, type='fixef')
}
\author{
Yian Guo, Yuting Yang, Xiaoman Zhai
}
