% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.plot.R
\name{pred.plot}
\alias{pred.plot}
\title{Generate prediction and observation plots.}
\usage{
pred.plot(mod, data, subject, time, response)
}
\arguments{
\item{mod}{a lmer model}

\item{data}{the dataset used in the object}

\item{subject}{name of the independent sampling unit}

\item{time}{name of the time variable associated with repeated measures}

\item{response}{name of the response variable}
}
\description{
This function plots predicted values and observed response values against time variable associated with repeated measures from a linear mixed effect model for visual comparison of model fit.
}
\examples{
data(sleepstudy)
mod<-lmer(Reaction ~ Days + (Days|Subject), data=sleepstudy)
pred.plot(mod, sleepstudy, subject='Subject', time='Days', response='Reaction')
}
\author{
Yian Guo, Yuting Yang, Xiaoman Zhai
}
