% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.plot.R
\name{resid.plot}
\alias{resid.plot}
\title{Plotting pearson residuals for linear mixed effect models with outliers denotation}
\usage{
\method{resid}{plot}(mod, denote = TRUE)
}
\arguments{
\item{mod}{a lmer model}

\item{denote}{gives the index with top 5 percent pearson type residuals}
}
\description{
This function plots the pearson type residuals against fitted values from a linear mixed effect model fitted by lmer function, with optional denotation of possible outliers.
}
\examples{
data(sleepstudy)
mod<-lmer(Reaction ~ Days + (Days|Subject), data=sleepstudy)
resid.plot(mod, denote=TRUE)
}
\author{
Yian Guo, Yuting Yang, Xiaoman Zhai
}
